requirements.txt
streamlit
pandas
plotly
psycopg2-binary
websocket-client
python-dotenv